fix: refactor issues
