+++ apps/container/src/App.vue
@@ -1,5 +1,5 @@
 <template>
-  <AppBar :showAppBar="confirmedUser">
+  <AppBar :showAppBar="confirmedUser ?? undefined">
     <v-main style="background-color: #f5f5f5">
       <v-container
         id="app-container"
@@ -21,7 +21,7 @@ import { ElMessage } from 'element-plus';
 
 import router from './router';
 
-import AppBar from '@/layouts/AppBar.vue';
+import { AppBar } from '@/layouts/AppBar';
 
 const useAuth = create(authStore);
 const { authenticate, user } = useAuth();
